Join params and types that deppend each other
