+++ src/types/manifest.ts
@@ -1,4 +1,3 @@
-import { architectures, chainDriversTypes } from "../params/index.js";
 import {
   SetupSchema,
   SetupTarget,
@@ -143,6 +142,13 @@ export type ChainDriverSpecs = {
 };
 
 export type ChainDriverType = (typeof chainDriversTypes)[number];
+export const chainDriversTypes = [
+  "bitcoin",
+  "ethereum",
+  "ethereum-beacon-chain",
+  "ethereum2-beacon-chain-prysm",
+  "monero",
+] as const;
 
 // HTTPS ports
 interface ExposableServiceManifestInfo {
@@ -170,3 +176,4 @@ export interface ManifestPaths {
 // Arch
 
 export type Architecture = (typeof architectures)[number];
+export const architectures = ["linux/amd64", "linux/arm64"] as const;
