+++ webpack.config.js
@@ -1,18 +1,6 @@
-import path from "path";
-import { fileURLToPath } from "url";
-
-const currentModulePath = fileURLToPath(import.meta.url);
-
 export default {
   entry: "./src/index.ts",
   mode: "production",
-  output: {
-    path: path.resolve(path.dirname(currentModulePath), "dist"),
-    filename: "index.js",
-    library: {
-      type: "module",
-    },
-  },
   resolve: {
     extensions: [".ts", ".tsx", ".js"],
     extensionAlias: {
@@ -22,7 +10,10 @@ export default {
     },
   },
   experiments: {
-    outputModule: true, // Enable the output module experiment
+    outputModule: true,
+  },
+  output: {
+    module: true,
   },
   module: {
     rules: [
